+++ backend/core/src/auth/dto/user.dto.ts
@@ -20,11 +20,11 @@ import { Match } from "../../shared/decorators/match.decorator"
 import { passwordRegex } from "../../shared/password-regex"
 
 export class UserDto extends OmitType(User, [
-  "isAdmin",
   "leasingAgentInListings",
   "passwordHash",
   "resetToken",
   "confirmationToken",
+  "roles",
 ] as const) {
   @Expose()
   @IsOptional()
@@ -35,10 +35,8 @@ export class UserDto extends OmitType(User, [
 }
 
 export class UserBasicDto extends OmitType(User, [
-  "isAdmin",
   "leasingAgentInListings",
   "passwordHash",
-  "roles",
   "confirmationToken",
   "resetToken",
 ] as const) {}
@@ -65,7 +63,6 @@ export class UserCreateDto extends OmitType(UserDto, [
   "createdAt",
   "updatedAt",
   "leasingAgentInListings",
-  "roles",
 ] as const) {
   @Expose()
   @IsString({ groups: [ValidationsGroupsEnum.default] })
@@ -98,7 +95,6 @@ export class UserUpdateDto extends OmitType(UserDto, [
   "createdAt",
   "updatedAt",
   "leasingAgentInListings",
-  "roles",
 ] as const) {
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })